Create helpers.rb for common methods
